fix: circular imports on DI system
