+++ kstreams/consts.py
@@ -1,3 +1,5 @@
+import enum
+
 APPLICATION_X_AVRO_BINARY = "application/x-avro-binary"
 APPLICATION_X_AVRO_JSON = "application/x-avro-json"
 APPLICATION_X_JSON_SCHEMA = "application/x-json-schema"
@@ -11,3 +13,15 @@ SERIALIZATION_TYPES = AVRO_SERIALIZATION_TYPES + (
     APPLICATION_X_JSON_SCHEMA,
     APPLICATION_JSON,
 )
+
+
+class UDFType(str, enum.Enum):
+    NO_TYPING = "NO_TYPING"
+    WITH_TYPING = "WITH_TYPING"
+
+
+class StreamErrorPolicy(str, enum.Enum):
+    RESTART = "RESTART"
+    STOP = "STOP"
+    STOP_ENGINE = "STOP_ENGINE"
+    STOP_APPLICATION = "STOP_APPLICATION"
